SECTIONAL ROLE VALIDATION!!!

+++ botCommands/administrative.py
@@ -20,6 +20,8 @@ WATERLOO_API_URL = "https://api.uwaterloo.ca/v2/directory/"
 redisClient = redis.Redis(host='localhost', port=6379, db=0)
 
 TOKEN = "NzA2Njc4Mzk2MzEwMjU3NzI1.Xq9v2A.iCXfvgwxz4fnmlrRUvTlA_JnSTA"
+section2List = ["saaliyan","a9ahluwa","yhahn","kalatras","d22an","n22arora","j24au","g4aujla","s3aulakh","mavolio","e2baek","x53bai","d22baker","nbeilis","j39bi","ebilaver","jbodner","a23bose","j24brar","j6braun","r6bui","gbylykba","achalakk","v5chaudh","ichellad","h596chen","ly23chen","h559chen","ncherish","jchik","jchitkar","skcho","kchoa","e25chu","nchunghu","m24coope","asdhiman","j3enrigh","derisogl","d24ferna","lfournie","n6franci","agabuniy","a57garg","mgionet","sgoodarz","c2gravel","m8guan","a324gupt","wharris","a29he","c55he","chenfrey","e44ho","rhoffman","p23hu","h338huan","l247huan","a73huang","a226jain","z242jian","h56jin","pkachhia","kkalathi","e2koh","k5kumara","jklkundn","k26le","j763lee","d267lee","k323lee","rlevesqu","a284li","r374li","k36liang","j352lu","b49lu","mlysenko","vmago","smanakta","j78marti","rhmayilv","a47mehta","d36mehta","a2mladen","d6moon","a27nadee","b42nguye","dnnnguye","b43nguye","m22niu","snuraniv","t5oliver","motchet","m332pate","v227pate","b36peng","bphu","npotdar","m98rahma","msraihaa","jrintjem","rrouhana","o3salem","apsalvad","s5santhi","hsayedal","tshahjah","s4shahri","r4sim","a553sing","a558sing","ll3smith","j225smit","kb2son","dsribala","tstauffe","a6su","ssubbara","m38syed","w29tam","c46tan","w4tao","s4thapa","ctraxler","etroci","a2vaseeh","j23vuong","d7wan","j23weng","t54wong","yy8wong","y657wu","j478wu","cy2xi","c7xiang","k233yang","j52yoon","i6zhang","cf3zhang","c624zhan","z963zhan"]
+
 
 
 # Used for daemon tasks, such as removing temporary membership and etc.
@@ -166,10 +168,10 @@ class Administrative(commands.Cog, name='Administrative'):
             messageAuthor = ctx.author
             watid = args[0]
 
-            if (redisClient.exists(str(messageAuthor) + ".request")):
+            if (redisClient.exists(str(messageAuthor) + ".request") or redisClient.exists(str(messageAuthor.id) + ".request")):
                 response = "<@" + str(
                     messageAuthor.id) + "> There is already a pending verification request for your WatID," \
-                                        "please use `!confirm <code>` or do `!cancelverification`"
+                                        " please use `!confirm <code>` or do `!cancelverification`"
                 await ctx.send(response)
                 return
             # Ask UW API for information
@@ -180,24 +182,27 @@ class Administrative(commands.Cog, name='Administrative'):
             if (apiResponse['data']['department'] != "ENG/Electrical and Computer"):
                 response = "<@" + str(
                     messageAuthor.id) + "> You are not an ECE student!" \
-                                        "Please manually validate by contacting" \
-                                        "the admin team. The admin team has been" \
-                                        "notified of this incident. <@&706658128409657366>"
+                                        " Please manually validate by contacting" \
+                                        " the admin team. The admin team has been" \
+                                        " notified of this incident. <@&706658128409657366>"
                 await ctx.send(response)
                 return
             if (len(apiResponse['data']['telephone_numbers']) > 0):
                 response = "<@" + str(
                     messageAuthor.id) + "> You are a faculty member, and faculty members" \
-                                        "require manual validation by an administrative team member." \
-                                        "Please contact the administration team by messaging them directly," \
-                                        "or send an email to k5kumara@uwaterloo.ca."
+                                        " require manual validation by an administrative team member." \
+                                        " Please contact the administration team by messaging them directly," \
+                                        " or send an email to k5kumara@uwaterloo.ca."
                 await ctx.send(response)
                 return;
-            if (redisClient.exists(str(messageAuthor) + ".verified")):
-                if (int(redisClient.get(str(messageAuthor) + ".verified")) == 1):
-                    response = "<@" + str(messageAuthor.id) + "> You have already been verified"
-                    await ctx.send(response)
-                    return
+            try:
+                if (redisClient.exists(str(messageAuthor) + ".verified") or redisClient.exists(str(messageAuthor.id) + ".verified")):
+                    if (int(redisClient.get(str(messageAuthor) + ".verified")) == 1 or int(redisClient.get(str(messageAuthor.id) + ".verified")) == 1):
+                        response = "<@" + str(messageAuthor.id) + "> You have already been verified"
+                        await ctx.send(response)
+                        return
+            except:
+                print("Lazy nullify error.")
             if (redisClient.exists(str(user_id))):
                 if (int(redisClient.get(str(user_id))) == 1):
                     response = "<@" + str(
@@ -206,13 +211,15 @@ class Administrative(commands.Cog, name='Administrative'):
                     return
 
             # Mark
-            redisClient.set(str(messageAuthor) + ".watid", user_id)
-            redisClient.set(str(messageAuthor) + ".verified", 0)
+            #redisClient.set(str(messageAuthor) + ".watid", user_id)
+            redisClient.set(str(messageAuthor.id) + ".watid", user_id)
+            redisClient.set(str(messageAuthor.id) + ".verified", 0)
             redisClient.set(str(messageAuthor) + ".name", name)
+            redisClient.set(str(messageAuthor.id) + ".name", name)
 
             # Generate random code
             code = random.randint(1000, 9999)
-            redisClient.set(str(messageAuthor), code)
+            redisClient.set(str(messageAuthor.id) + ".code", code)
 
             mailMessage = Mail(
                 from_email='verification@kaveenk.com',
@@ -229,17 +236,17 @@ class Administrative(commands.Cog, name='Administrative'):
 
             response = "<@" + str(
                 messageAuthor.id) + "> I sent a verification code to " + email + ". Find the code" \
-                                                                                 "in your email and type `!confirm <code>` in discord to verify" \
-                                                                                 "your account. Please check your spam and junk folders."
-            redisClient.set(str(messageAuthor) + ".request", 1)
+                                                                                 " in your email and type `!confirm <code>` in discord to verify" \
+                                                                                 " your account. Please check your spam and junk folders."
+            redisClient.set(str(messageAuthor.id) + ".request", 1)
 
             await ctx.send(response)
         except Exception as e:
             print(e)
             response = "<@" + str(
                 messageAuthor.id) + "> No WatID provided or invalid watID, please use `!verify <watid>`." \
-                                    "Your WatID is the username in your original email, for example, in " \
-                                    "k5kumara@edu.uwaterloo.ca, the watID is k5kumara."
+                                    " Your WatID is the username in your original email, for example, in " \
+                                    " k5kumara@edu.uwaterloo.ca, the watID is k5kumara."
             await ctx.send(response)
 
     @commands.command()
@@ -249,28 +256,29 @@ class Administrative(commands.Cog, name='Administrative'):
 
             code = args[0]
 
-            if (redisClient.exists(str(messageAuthor) + ".request")):
+            if (redisClient.exists(str(messageAuthor.id) + ".request")):
 
-                if (int(code) == int(redisClient.get(str(messageAuthor)))):
+                if (int(code) == int(redisClient.get(str(messageAuthor.id)+".code"))):
                     response = "<@" + str(messageAuthor.id) + "> You were successfully verified."
 
                     await ctx.send(response)
 
-                    nickname = redisClient.get(str(messageAuthor) + ".name")
+                    nickname = redisClient.get(str(messageAuthor.id) + ".name")
 
                     await messageAuthor.edit(nick=str(nickname.decode('utf-8')))
 
                     # Mark user and WatID as verified
-                    redisClient.set(str(messageAuthor) + ".verified", 1)
-                    redisClient.set(str(redisClient.get(str(messageAuthor) + ".watid").decode('utf-8')), 1)
-                    redisClient.delete(str(messageAuthor) + ".request")
+                    redisClient.set(str(messageAuthor.id) + ".verified", 1)
+                    #redisClient.set(str(redisClient.get(str(messageAuthor) + ".watid").decode('utf-8')), 1)
+                    redisClient.set(str(redisClient.get(str(messageAuthor.id) + ".watid").decode('utf-8')), 1)
+
+                    if (redisClient.exists(str(messageAuthor.id))): redisClient.delete(str(messageAuthor.id) + ".request")
+                    if (redisClient.exists(str(messageAuthor))): redisClient.delete(str(messageAuthor) + ".request")
                     # 706966831268626464
                     verifiedRole = discord.utils.get(ctx.guild.roles, name="Verified")
                     unverifiedRole = discord.utils.get(ctx.guild.roles, name="Unverified")
                     await messageAuthor.add_roles(verifiedRole)
                     await messageAuthor.remove_roles(unverifiedRole)
-
-
 
                 else:
                     response = "<@" + str(messageAuthor.id) + "> Invalid verification code."
@@ -292,7 +300,8 @@ class Administrative(commands.Cog, name='Administrative'):
         messageAuthor = ctx.author
 
         # 706966831268626464
-        if (redisClient.exists(str(messageAuthor) + ".request")):
+        if (redisClient.exists(str(messageAuthor.id) + ".request")):
+            redisClient.delete(str(messageAuthor.id))
             response = "<@" + str(
                 messageAuthor.id) + "> Cancelled your on-going verification, please try again with `!verify <watid>`"
             await ctx.send(response)
@@ -311,20 +320,24 @@ class Administrative(commands.Cog, name='Administrative'):
                 allowed = True
         if (allowed):
             try:
+
                 selection = args[0]
                 if (selection == "user"):
                     user = ctx.message.mentions[0]
-                    watid = redisClient.get(str(user) + ".watid").decode('utf-8')
+                    watid = redisClient.get(str(user.id) + ".watid").decode('utf-8')
                     redisClient.delete(watid)
                     await ctx.send("Unmarked WatID " + watid)
                     redisClient.delete(str(user) + ".watid")
+                    redisClient.delete(str(user.id) + ".watid")
                     await ctx.send("Purged WatID")
+                    redisClient.delete(str(user.id) + ".verified")
                     redisClient.delete(str(user) + ".verified")
                     await ctx.send("Purged verified status")
                     redisClient.delete(str(user) + ".name")
+                    redisClient.delete(str(user.id) + ".name")
                     await ctx.send("Purged legal name")
                     redisClient.delete(str(messageAuthor))
-                    redisClient.delete(str(user) + ".request")
+                    redisClient.delete(str(user.id) + ".request")
                     await ctx.send("Purged request status")
                     await ctx.send("Purged user from database successfully.")
 
@@ -335,8 +348,8 @@ class Administrative(commands.Cog, name='Administrative'):
                 else:
                     await ctx.send("<@" + str(
                         messageAuthor.id) + "> Invalid selection! You can choose to devalidate a user or a WatID.")
-            except:
-                print("<@+" + str(
+            except Exception as e:
+                ctx.send("<@" + str(
                     messageAuthor.id) + "> Invalid syntax or selection: `!devalidate <select 'user' or 'watid'> <value>`")
 
     @commands.command()
@@ -366,7 +379,8 @@ class Administrative(commands.Cog, name='Administrative'):
                     await ctx.send("Invalid WatID: " + watid)
                     return
 
-                redisClient.set(str(user) + ".watid", watid)
+                #redisClient.set(str(user) + ".watid", watid)
+                redisClient.set(str(user.id) + ".watid", watid)
                 await ctx.send("WatID " + watid + " has been validated and correlated to <@" + str(user.id) + ">")
                 if ("Verified" in ranks):
                     redisClient.set(str(user) + ".verified", 1)
@@ -375,7 +389,7 @@ class Administrative(commands.Cog, name='Administrative'):
                 await user.edit(nick=name)
                 await ctx.send(
                     "Name " + name + " has been validated and correlated to <@" + str(user.id) + ">")
-                redisClient.set(str(redisClient.get(str(messageAuthor) + ".watid").decode('utf-8')), 1)
+                redisClient.set(watid, 1)
                 await ctx.send(
                     "The WatID " + watid + " has been marked for no further verifications.")
 
@@ -408,7 +422,6 @@ class Administrative(commands.Cog, name='Administrative'):
                 await ctx.send("All tasks completed successfully")
             except Exception as e:
                 print(str(e))
-                print('t4')
                 await ctx.send("<@" + str(
                     messageAuthor.id) + "> You have entered invalid syntax, or the user you are trying to correlate is invalid. `!correlate <USER MENTION> <WatID>`")
 
@@ -431,7 +444,7 @@ class Administrative(commands.Cog, name='Administrative'):
 
                     # Find user's discord tag
                     for member in ctx.message.mentions:
-                        discordID = str(member)
+                        discordID = str(member.id)
                         watid = redisClient.get(discordID + ".watid").decode('utf-8')
                         break
                 apiResponse = requests.get(WATERLOO_API_URL + watid + ".json?key=" + WATERLOO_API_KEY).json()
@@ -491,6 +504,7 @@ class Administrative(commands.Cog, name='Administrative'):
 
     @commands.command()
     async def validateroles(self, ctx):
+        adminChannel = discord.utils.get(ctx.author.guild.channels, id=716954090495541248)
 
         messageAuthor = ctx.author
 
@@ -499,20 +513,43 @@ class Administrative(commands.Cog, name='Administrative'):
             if role.name == 'Admin':
                 allowed = True
         if (allowed):
+            section1Role = discord.utils.get(ctx.message.guild.roles, name="Section 1")
+            section2Role = discord.utils.get(ctx.message.guild.roles, name="Section 2")
             verifiedRole = discord.utils.get(ctx.message.guild.roles, name="Verified")
-            unverifiedRole = discord.utils.get(ctx.message.guild.roles, name="Unverified")
-            adminRole = discord.utils.get(ctx.message.guild.roles, name="Admin")
+            guestRole = discord.utils.get(ctx.message.guild.roles, name="Guest")
             teachingRole = discord.utils.get(ctx.message.guild.roles, name="Teaching Staff")
 
-            memberList = ctx.message.guild.members
-            for member in memberList:
-                if (verifiedRole in member.roles and unverifiedRole in member.roles):
-                    await ctx.send("Removed unverified role from " + member.name)
-                    await member.remove_roles(unverifiedRole)
-                elif (
-                        verifiedRole not in member.roles and unverifiedRole not in member.roles and adminRole not in member.roles and teachingRole not in member.roles):
-                    await ctx.send("Added unverified role to " + member.name)
-                    await member.add_roles(unverifiedRole)
+
+            for member in ctx.author.guild.members:
+                if (teachingRole in member.roles or guestRole in member.roles or verifiedRole not in member.roles):
+                    continue
+
+                try:
+                    if (redisClient.exists(str(member.id)+".watid") and not redisClient.exists(str(member.id)+".rolevalidated")):
+                        await adminChannel.send("Analyzing user <@"+str(member.id)+">")
+                        watID = redisClient.get(str(member.id) + ".watid").decode("utf-8")
+                        await adminChannel.send("The WatID for user <@" + str(member.id) + "> is "+watID)
+
+                        await member.remove_roles(section1Role)
+                        await member.remove_roles(section2Role)
+                        if (watID in section2List):
+                            await member.add_roles(section2Role)
+                            await adminChannel.send("Added the Section 2 Role to <@"+str(member.id)+">.")
+                        else:
+                            await member.add_roles(section1Role)
+                            await adminChannel.send("Added the Section 1 Role to <@" + str(member.id) + ">.")
+                        redisClient.set(str(member.id)+".rolevalidated","true")
+
+                    else:
+                        await adminChannel.send("<@&706658128409657366> There was no WatID for: <@" + str(
+                            member.id) + "> please investigate.")
+
+                except:
+                    await adminChannel.send("<@&706658128409657366> There was an error retriving the WatID for: <@"+str(member.id)+"> please investigate.")
+
+
+
+
             await ctx.send("All role validations completed successfully.")
 
     @commands.command()
